requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
astroid==3.3.10
black==25.1.0
certifi==2025.4.26
cffi==1.17.1
click==8.2.1
colorama==0.4.6
dill==0.4.0
Django==5.2.2
django-minio-backend==3.8.0
djangorestframework==3.16.0
drf-yasg==1.21.10
inflection==0.5.1
isort==6.0.1
mccabe==0.7.0
minio==7.2.15
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
pycryptodome==3.23.0
pylint==3.3.7
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
tomlkit==0.13.3
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
